+++ .htmlhintrc
@@ -1,0 +1,3 @@
+{
+  "doctype-first": false
+}
