+++ .stylelintrc.json
@@ -1,0 +1,1 @@
+{ "extends": ["stylelint-config-standard"] }
