+++ html/chunks/rules_modals.html
@@ -1,4 +1,3 @@
-<!-- pop up explanations. These are modals that can be used across pages, if imported -->
 <div id="percentage-modal" class="modal fade" tabindex="-1" role="dialog" aria-labelledby="" aria-hidden="true">
     <div class="modal-dialog modal-dialog-centered" role="document">
         <div class="modal-content">
