+++ index.html
@@ -16,6 +16,7 @@
     <title>Riocontra Translator | Traduttore dall'italiano al riocontra</title>
   </head>
   
+  <body>
   <nav class="navbar navbar-dark bg-primary">
     <div class="container">
       <a class="navbar-brand on-primary-text" href="#">Riocontra Translator</a>
