+++ package.json
@@ -14,6 +14,9 @@
   "devDependencies": {
     "eslint": "^8.38.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.27.5"
+    "eslint-plugin-import": "^2.27.5",
+    "htmlhint": "^1.1.4",
+    "stylelint": "^15.6.0",
+    "stylelint-config-standard": "^33.0.0"
   }
 }
